+++ src/main/kotlin/io/openfuture/chain/domain/transaction/PendingTransactionRequest.kt
@@ -1,9 +1,0 @@
-package io.openfuture.chain.domain.transaction
-
-open class PendingTransactionRequest(
-    val amount: Int,
-    val timestamp: Long,
-    val recipientKey: String,
-    val senderKey: String,
-    val signature: String
-)