+++ src/main/kotlin/io/openfuture/chain/service/DefaultTransactionService.kt
@@ -1,19 +1,18 @@
 package io.openfuture.chain.service
 
-import io.openfuture.chain.domain.transaction.TransactionRequest
 import io.openfuture.chain.entity.Transaction
 import io.openfuture.chain.repository.TransactionRepository
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 
 @Service
 class DefaultTransactionService(
-    private val transactionRepository: TransactionRepository,
-    private val blockService: DefaultBlockService
-): TransactionService {
+    private val repository: TransactionRepository
+) : TransactionService {
 
-    override fun save(request: TransactionRequest) {
-        val block = blockService.get(request.blockId)
-        transactionRepository.save(Transaction.of(block, request))
+    @Transactional
+    override fun save(transaction: Transaction): Transaction {
+        return repository.save(transaction)
     }
 
 }