+++ src/main/resources/db/migration/V1_4__table_wallets.sql
@@ -1,0 +1,5 @@
+CREATE TABLE wallets (
+  id            INTEGER PRIMARY KEY,
+  address       VARCHAR NOT NULL,
+  balance       DOUBLE
+);